requirements/base.txt
cchardet
aiodns
aiohttp
aiohttp-jinja2
python-dotenv
sqlalchemy
asyncpg
aiofiles
asyncio
psycopg2-binary
celery